"About Me" Greatly Expanded; Removed Dead Project from "Projects"
Added content from resume onto the "About Me" tab. Also did some
formatting to make things look nice.

Remove SCFbackup from "Projects" tab for now, as it is kind of dead in
the water. There's a lot more work that needs to be done to it, and I
would like to return to it. However, for now it is indefinitely
suspended.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>